+++ app/Http/Controllers/TagsController.php
@@ -95,6 +95,11 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if($tag->posts->count() > 0){
+            session()->flash('error', 'This tag cannot be deleted as it has some posts');
+            return redirect()->back();
+        }
+        
         $tag->delete();
 
         session()->flash('success', 'Tag deleted successfully');
